WL-910 Upgraded the en_GB localisation of gradebook.
I pretty much did this from scratch, i.e. I took the default Messages.properties
file and worked through each property. I was looking for references to grading,
gradebook, grading, etc in particular. I then compared this to the previous
revision and was able to tidy things up a bit (including spotting a few
duplicates!) in order to minimise the overall size of the diff.